이미지 캐싱 로직 분리
