+++ WeatherForecast/WeatherForecast/Model/ImageLoader.swift
@@ -1,0 +1,34 @@
+//
+//  ImageLoader.swift
+//  WeatherForecast
+//
+//  Created by Daegeon Choi on 2024/02/07.
+//
+
+import Foundation
+import UIKit
+
+final class ImageLoader {
+    
+    static let shared: ImageLoader = ImageLoader()
+    
+    let imageChache: NSCache<NSString, UIImage> = NSCache()
+    
+    func fetchImage(urlString: String, _ completion: @escaping (UIImage) -> Void) {
+        if let image = imageChache.object(forKey: urlString as NSString) {
+            completion(image)
+            return
+        }
+        
+        Task {
+            guard let url: URL = URL(string: urlString),
+                  let (data, _) = try? await URLSession.shared.data(from: url),
+                  let image: UIImage = UIImage(data: data) else {
+                return
+            }
+            
+            imageChache.setObject(image, forKey: urlString as NSString)
+            completion(image)
+        }
+    }
+}
